refactor(lotto-checker): split draw_history_screen widgets
- What: extract DrawCard, DateChip, TierRow into widgets/draw_card.dart;
  extract EmptyView, ErrorView into widgets/draw_history_states.dart.
  EmptyView now accepts optional title/subtitle (default unchanged).
- Why: prep for F2 filter feature — file was at 250-line ceiling.
  Split keeps each file well under the limit and gives the filter
  state UI room to land cleanly.
- Impact: draw_history_screen.dart 250 → 56 lines. No behavior change.
  Existing 6 widget tests still pass.
